chore: move fallback to host directive
