+++ projects/kit/components/line-clamp/index.ts
@@ -1,3 +1,4 @@
 export * from './line-clamp.component';
 export * from './line-clamp.options';
 export * from './line-clamp-box.component';
+export * from './line-clamp-fallback.directive';
